set minimum window size to 1x1
This commit changes the minimum window size to 1x1 to allow the user to
freely resize the window. Without this patch, windows would be at least
as wide as the URL textbox, which is bad because when on a page with a
short URL and hovering a link with a very long URL, the window size
changes apruptly.

Signed-off-by: Gregor Best <farhaven@googlemail.com>