Implement locking on revive() logic.
When writing out the file for the use of reviving the last URL (which is
useful if Vimprobable is being used inside tabbed, for example), ensure we
at least try and not clobber the same file if two or more instances of
Vimprobable try and write to the same file.

Note that this isn't so much about ensuring searialisation as it is about
crudely disallowing file corruption when writing/reading from the URL file.