Implement persistent-cookie support.
Previously, Vimprobable only defined one cookie-store per running instance;
effectively, this meant that cookies set by one instance of Vimprobable
could never see the newer cookies, because the cookiejar was only ever
queried once; when it loaded.

Introduce parsing of the header requests we receive so that we can
interrogate cookies per-request.  This will give the allusion of "shared"
cookies.