Fix "map" processing in config files.
Trying to calculate the size of an extern array at runtime is fatal -- it's
also going to either evaluate to zero (if passed by pointer) or fail at
compilation, because sizeof is compile-time only.

Therefore, give the commands[] array a pre-calculated size -- which can
then be used properly when trying to set any "map" lines encountered in
Vimprobable's config file.