Use "changed" signal to update cookie jars.
Go back to using libsoup's "changed" signal handling on the file-store for
the on disk cookie-jar, and update our session-store as it's updated.

Request headers from URLs which contain cookies go straight into the file
cookie jar on disk.

This should help reduce the overhead of needing to read/write to the entire
file cookie store on disk each time.