Allow for per-download paths to be set
This introduces a Download struct for handling the meta data surrounding
downloads.  In the case of promptdownload being set to true, we must
maintain the state of the download before we prompt the user for the path.

This also extends logically to "automatic" downloads, that is, with
promptdownload set to false.