Numbers wouldn't open hints in new window. Fixed.
There was a problem with relying on followTarget to see whether vimprobable was
in hint mode. This was caused because within script(), inputbox_changed_cb would
be called, and inputbox_changed_cb memset followTarget, so hints would always be
followed in same window when using numbers. Now, followTarget is only memset
before "current" or "new" is assigned to it, and mode is set to ModeHints to
signify vimprobable is hints mode.

Fixed bug where numbers won't open background hint in background