Implemented removing of hints if a number is selected.
I adapted the behaviour to the vimperator plugin, which does remove
hints if a number was entered, but without regeneration of the numberes
hint spans. Also implemented cyclic walk via tab or shift tab through th
hints - if reached the begin or end go to the other end wit hthe focus.